+++ app/src/main/AndroidManifest.xml
@@ -24,8 +24,8 @@
         <activity android:name=".ui.login.LoginActivity"/>
         <activity android:name=".ui.menu.MainMenuActivity" />
         <activity android:name=".ui.pilihLawan.PilihLawan"/>
-        <activity android:name=".playGame.MainGamePlayer" />
-        <activity android:name=".playGame.MainGameComputer"/>
+        <activity android:name="com.teamacodechallenge6.ui.playGame.MainGamePlayer" />
+        <activity android:name="com.teamacodechallenge6.ui.playGame.MainGameComputer"/>
         <activity android:name="com.teamacodechallenge6.ui.profileteman.ProfileTeman"/>
     </application>
