+++ app/src/main/java/com/teamacodechallenge6/App.kt
@@ -2,12 +2,12 @@ package com.teamacodechallenge6
 
 import android.app.Application
 import android.content.Context
-import com.teamacodechallenge6.database.TemanDatabase
+import com.teamacodechallenge6.data.database.TemanDatabase
 
 class App: Application() {
     companion object{
         var context: Context? = null
-        var mDB:TemanDatabase?=null
+        var mDB: TemanDatabase?=null
     }
 
     override fun onCreate() {
