+++ app/src/main/java/com/teamacodechallenge6/ui/login/LoginPresenterImp.kt
@@ -3,7 +3,7 @@ package com.teamacodechallenge6.ui.login
 import com.teamacodechallenge6.App.Companion.context
 import com.teamacodechallenge6.App.Companion.mDB
 import com.teamacodechallenge6.data.local.SharedPref
-import com.teamacodechallenge6.database.TemanDatabase
+import com.teamacodechallenge6.data.database.TemanDatabase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
