+++ app/src/main/java/com/teamacodechallenge6/ui/menu/MainMenuActivity.kt
@@ -6,7 +6,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.google.android.material.snackbar.Snackbar
 import com.teamacodechallenge6.R
-import com.teamacodechallenge6.playGame.MainGameComputer
+import com.teamacodechallenge6.ui.playGame.MainGameComputer
 import com.teamacodechallenge6.ui.login.LoginActivity
 import com.teamacodechallenge6.ui.pilihLawan.PilihLawan
 import com.teamacodechallenge6.ui.profileteman.ProfileTeman
@@ -14,6 +14,7 @@ import kotlinx.android.synthetic.main.activity_main.*
 
 class MainMenuActivity : AppCompatActivity(), MainMenuView {
     private lateinit var presenter: MainMenuPresenter
+    private val username= mutableListOf<String>()
     val parent: ConstraintLayout by lazy { findViewById(R.id.main) }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -44,6 +45,7 @@ class MainMenuActivity : AppCompatActivity(), MainMenuView {
         snackbar.setAction("Tutup"){
             snackbar.dismiss()
         }.show()
+        username.add(msg)
         tvPlayer.text=msg
         tvPlayer2.text=msg
     }
