+++ app/src/main/java/com/teamacodechallenge6/ui/pilihLawan/DaftarPemain.kt
@@ -4,7 +4,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.teamacodechallenge6.R
-import com.teamacodechallenge6.database.TemanDatabase
+import com.teamacodechallenge6.data.database.TemanDatabase
 import kotlinx.android.synthetic.main.activity_pilih_lawan.*
 import kotlinx.android.synthetic.main.activity_profile_teman.*
 import kotlinx.android.synthetic.main.activity_profile_teman.recyclerView
