+++ app/src/main/java/com/teamacodechallenge6/ui/pilihLawan/PilihLawanPresenter.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import androidx.recyclerview.widget.RecyclerView
 
 interface PilihLawanPresenter {
-//    fun signUp(username: String, password: String, email: String)
     fun showList(recyclerView: RecyclerView, context: Context)
     fun destroyDB()
 }
