+++ app/src/main/java/com/teamacodechallenge6/ui/playGame/Gameplay.kt
@@ -1,4 +1,4 @@
-package com.teamacodechallenge6.playGame
+package com.teamacodechallenge6.ui.playGame
 
 data class Gameplay(val dataPlayer1: String, val dataPlayer2:String, val modeGame:String)
