+++ app/src/main/java/com/teamacodechallenge6/ui/playGame/MainGamePresenter.kt
@@ -1,4 +1,4 @@
-package com.teamacodechallenge6.playGame
+package com.teamacodechallenge6.ui.playGame
 
 interface MainGamePresenter {
     fun resultEnemy(resultEnemy:String)
