+++ app/src/main/java/com/teamacodechallenge6/ui/profileteman/TemanPresenter.kt
@@ -2,7 +2,7 @@ package com.teamacodechallenge6.ui.profileteman
 
 import android.content.Context
 import androidx.recyclerview.widget.RecyclerView
-import com.teamacodechallenge6.database.Teman
+import com.teamacodechallenge6.data.database.Teman
 
 interface TemanPresenter {
     fun playerName()
@@ -10,5 +10,6 @@ interface TemanPresenter {
     fun listTeman(recyclerView: RecyclerView, context: Context)
     fun deleteTeman(list:List<Teman>, position: Int)
     fun editTeman(list:List<Teman>, position: Int)
+
     fun destroyDB()
 }
